refactor: split settings into two components
